Move isLoanLiquidated and determineLoanStatus to loans utils
